LPS-50752 move upgrade process into service
