+++ modules/apps/social/social-networking-service/src/com/liferay/socialnetworking/upgrade/v1_0_0/UpgradeNamespace.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.socialnetworking.hook.upgrade.v1_0_0;
+package com.liferay.socialnetworking.upgrade.v1_0_0;
 
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
@@ -20,9 +20,9 @@ import com.liferay.portal.kernel.upgrade.UpgradeProcess;
 import com.liferay.portal.kernel.upgrade.util.UpgradeTable;
 import com.liferay.portal.kernel.upgrade.util.UpgradeTableFactoryUtil;
 import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.socialnetworking.hook.upgrade.v1_0_0.util.MeetupsEntryTable;
-import com.liferay.socialnetworking.hook.upgrade.v1_0_0.util.MeetupsRegistrationTable;
-import com.liferay.socialnetworking.hook.upgrade.v1_0_0.util.WallEntryTable;
+import com.liferay.socialnetworking.upgrade.v1_0_0.util.MeetupsEntryTable;
+import com.liferay.socialnetworking.upgrade.v1_0_0.util.MeetupsRegistrationTable;
+import com.liferay.socialnetworking.upgrade.v1_0_0.util.WallEntryTable;
 
 /**
  * @author Julio Camarero
