+++ modules/apps/social/social-networking-service/src/com/liferay/socialnetworking/upgrade/v1_0_0/util/MeetupsRegistrationTable.java
@@ -12,30 +12,31 @@
  * details.
  */
 
-package com.liferay.socialnetworking.hook.upgrade.v1_0_0.util;
+package com.liferay.socialnetworking.upgrade.v1_0_0.util;
 
 import java.sql.Types;
 
 /**
  * @author Brian Wing Shun Chan
  */
-public class WallEntryTable {
+public class MeetupsRegistrationTable {
 
 	public static final Object[][] TABLE_COLUMNS = {
-		{"wallEntryId", new Integer(Types.BIGINT)},
-		{"groupId", new Integer(Types.BIGINT)},
+		{"meetupsRegistrationId", new Integer(Types.BIGINT)},
 		{"companyId", new Integer(Types.BIGINT)},
 		{"userId", new Integer(Types.BIGINT)},
 		{"userName", new Integer(Types.VARCHAR)},
 		{"createDate", new Integer(Types.TIMESTAMP)},
 		{"modifiedDate", new Integer(Types.TIMESTAMP)},
+		{"meetupsEntryId", new Integer(Types.BIGINT)},
+		{"status", new Integer(Types.INTEGER)},
 		{"comments", new Integer(Types.VARCHAR)}
 	};
 
-	public static final String TABLE_NAME = "SN_WallEntry";
+	public static final String TABLE_NAME = "SN_MeetupsRegistration";
 
-	public static final String TABLE_SQL_CREATE = "create table SN_WallEntry (wallEntryId LONG not null primary key,groupId LONG,companyId LONG,userId LONG,userName VARCHAR(75) null,createDate DATE null,modifiedDate DATE null,comments STRING null)";
+	public static final String TABLE_SQL_CREATE = "create table SN_MeetupsRegistration (meetupsRegistrationId LONG not null primary key,companyId LONG,userId LONG,userName VARCHAR(75) null,createDate DATE null,modifiedDate DATE null,meetupsEntryId LONG,status INTEGER,comments STRING null)";
 
-	public static final String TABLE_SQL_DROP = "drop table SN_WallEntry";
+	public static final String TABLE_SQL_DROP = "drop table SN_MeetupsRegistration";
 
 }
